pyproject.toml
[tool.poetry]
name = "backend-template"
version = "0.1.0"
description = ""
authors = ["peileon2 <peileon2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.2.1"
cryptography = "^42.0.7"
requests = "^2.32.3"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
passlib = "^1.7.4"
emails = "0.6"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
sentry-sdk = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
